+++ lib/src/routing/routes.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_application_1/src/features/auth/screens/forgetPswrd/forget_pswrd_mail.dart';
 import 'package:flutter_application_1/src/features/users/business_owner/owner_homepage.dart';
 import 'package:flutter_application_1/src/features/users/customer/cust_homepage.dart';
-import 'package:flutter_application_1/src/features/users/customer/cust_profile.dart';
-import 'package:flutter_application_1/src/features/users/customer/edit_profie.dart';
+import 'package:flutter_application_1/src/features/users/customer/profile_page/cust_profile.dart';
+import 'package:flutter_application_1/src/features/users/customer/profile_page/edit_profie.dart';
 import 'package:flutter_application_1/src/features/users/customer/test_read_data.dart';
 import 'package:flutter_application_1/src/features/users/deliveryman/delivery_homepage.dart';
 import 'package:flutter_application_1/src/features/auth/screens/login/login_page.dart';
